requirements.txt
absl-py==0.5.0
annoy==1.13.0
astor==0.7.1
awscli==1.16.20
bleach==1.5.0
boto3==1.7.39
botocore==1.10.84
colorama==0.3.9
coverage==4.5.1
cycler==0.10.0
docutils==0.14
gast==0.2.0
geographiclib==1.49
geopy==1.17.0
grpcio==1.15.0
h5py==2.8.0
html5lib==0.9999999
jieba==0.39
jmespath==0.9.3
kafka-python==1.4.3
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.0
nose==1.3.7
numpy==1.15.2
olefile==0.46
opencv-python==3.4.3.18
pandas==0.23.4
parameterized==0.6.1
Pillow==4.1.1
protobuf==3.6.1
psycopg2
psycopg2-binary
pyasn1==0.4.4
Pygments==2.2.0
pyparsing==2.2.1
python-dateutil==2.7.3
pytz==2018.5
rsa==3.4.2
requests
s3transfer==0.1.13
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
sparkdl==0.2.2
tensorboard==1.10.0
tensorflow==1.10.0
tensorflowonspark==1.3.4
tensorframes==0.2.9
termcolor==1.1.0
tqdm==4.26.0
urllib3==1.23
Werkzeug==0.14.1
